[APT009] - Change type deci to float

+++ Apteczka/Apteczka.Data/DTO/APTLocation.cs
@@ -24,8 +24,8 @@ namespace Apteczka.Data.DTO
         public long APTPharmacyId { get; set; }
         public string City { get; set; }
         public string Street { get; set; }
-        public Nullable<decimal> Latitude { get; set; }
-        public Nullable<decimal> Longitude { get; set; }
+        public Nullable<double> Latitude { get; set; }
+        public Nullable<double> Longitude { get; set; }
     
         public virtual APTPharmacy APTPharmacy { get; set; }
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
